♻️ refactor: extract sync.helpers into separate dirs
This makes the sync code more readable by assigning functions to their respective features:
- import: getting events from Gcal -> Compass
- init: setting up the sync
- log: logging what happened during a sync
- watch: watching for events in gcal
