+++ packages/backend/src/auth/services/compass.auth.service.ts
@@ -2,7 +2,7 @@ import Session from "supertokens-node/recipe/session";
 import { SyncError } from "@backend/common/constants/error.constants";
 import { error } from "@backend/common/errors/handlers/error.handler";
 import { getSync } from "@backend/sync/util/sync.queries";
-import { canDoIncrementalSync } from "@backend/sync/util/sync.utils";
+import { canDoIncrementalSync } from "@backend/sync/util/sync.util";
 import { findCompassUserBy } from "@backend/user/queries/user.queries";
 
 class CompassAuthService {
