+++ packages/backend/src/sync/services/log/sync.logger.ts
@@ -23,7 +23,7 @@ export const logSyncOperation = async (
   const timestamp = dayjs().format("YYYY-MM-DD_HH-mm-ss");
   const logDir = path.join(process.cwd(), "logs", "sync");
   const filename = `sync_${userId}_${timestamp}.json`;
-  let fullPath = path.resolve(logDir, filename);
+  const fullPath = path.resolve(logDir, filename);
 
   // Ensure the fullPath is within the logDir
   if (!fullPath.startsWith(logDir)) {
